expose pkg functionality in the root namespace some ruff tidying up
